Moved Boardgame model to boardgames module
