+++ boardgames/tests/test_views.py
@@ -6,8 +6,8 @@ from django.views.generic.list import ListView
 from utils.testing.view_test_utils import ViewValidityMixin
 from utils.views import SearchFormMixin
 
-from inventory.models import Ownership, BoardGame
-from inventory.views import *
+from boardgames.models import Ownership, BoardGame
+from boardgames.views import *
 
 
 class TestBoardGameView(ViewValidityMixin, TestCase):
