+++ inventory/tests/test_models.py
@@ -6,7 +6,7 @@ from django.utils.text import slugify
 
 
 from membership_file.models import Member
-from inventory.models import Item, BoardGame, Ownership, valid_item_class_ids, ItemManager,\
+from inventory.models import Item, Ownership, valid_item_class_ids, ItemManager,\
     get_item_image_upload_path, MiscellaneousItem
